Crate strymon_job

The Strymon client-library. In order to register a Timely Dataflow job with Strymon, users are expected to link against this library and eventually invoke strymon_job::execute.

In addition, this library also provides the operaters needed to use the publish-subscribe mechanism.

Examples

extern crate strymon_job;
extern crate timely;

use timely::dataflow::operators::ToStream;
use strymon_job::operators::publish::Partition;

fn main() {
    strymon_job::execute(|worker, coord| {
        worker.dataflow::<u64, _, _>(|scope| {
            let stream = (0..1000).to_stream(scope);
            coord.publish("numbers", &stream, Partition::Merge)
                .expect("failed to publish topic");
        });
    }).unwrap();
}

Modules

operators
Types used by Strymon-specific operators such as publish and subscribe.

Structs

Coordinator
Handle to communicate with the Strymon coordinator.

Functions

execute
Executes a Timely dataflow within this Strymon job.