Function strymon_job::execute

pub fn execute<T, F>(func: F) -> Result<WorkerGuards<T>, String> where
    T: Send + 'static,
    F: Fn(&mut Root<Allocator>, Coordinator) -> T,
    F: Send + Sync + 'static, 

Executes a Timely dataflow within this Strymon job.

This function requires that the calling process has been spawned by an executor. Upon successful registration with the Strymon coordinator, the closure func is invoked for each requested worker hosted by the current process.

This function intentionally mirrors timely::execute, with the difference that the worker configuration is provided by the parent executor and that the running worker gains the ability to talk to the coordinator.