Struct strymon_job::operators::service::Client

pub struct Client<N: Name> { /* fields omitted */ }

Client handle for sending requests to a announced service.

Will deregister the announced service from the catalog when dropped.

Methods

impl<N: Name> Client<N>

pub fn request<R: Request<N>>(&self, r: &R) -> Response<N, R>

Asynchronously sends out a request to the remote peer.

Returns a future for the pending response. The next request can be submitted without having to wait for the previous response to arrive.