Enum strymon_job::operators::subscribe::SubscriptionEvent

pub enum SubscriptionEvent<T, D> {
    Data(T, Vec<D>),
    FrontierUpdate,
}

Data or frontier events emitted by a subscription.

Variants
Data(T, Vec<D>)

The Data marks the arrival of a data batch which has been published.

FrontierUpdate

The FrontierUpdate event indicates that the upstream frontier has changed. The new frontier can be inspected using the frontier() method on the subscription handle.